Rework of points and updated CUDA version at 12.6

+++ mudock/include/mudock/type_alias.hpp
@@ -2,6 +2,6 @@
 
 namespace mudock {
 
-  using fp_type = float;
+  using fp_type = double;
 
 } // namespace mudock
